chore: move type and interface declarations to a separate file
